move TaskParameters to core module
