+++ server/odc-core/src/main/java/com/oceanbase/odc/core/flow/model/TaskParameters.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.oceanbase.odc.service.flow.model;
+package com.oceanbase.odc.core.flow.model;
 
 /**
  * {@link TaskParameters}
